Added a new version of the stringbuilder class.
This version of the stringbuilder class is a lot faster than the old version.
It use a dictionary to keep all the single strings and concatenate it with a join if you call the toString() method.
In a short benchmark the new version is more than twice as fast as the old version.

stringbuilder.asp Version 3.0.0

Written by Markus Paeschke <http://www.mpaeschke.de> @ November 2012